Move Java plugin to a separate package.
