+++ unified-prototype/unified-plugin/plugin-jvm/build.gradle.kts
@@ -16,7 +16,7 @@ gradlePlugin {
         }
         create("java-application") {
             id = "org.gradle.experimental.java-application"
-            implementationClass = "org.gradle.api.experimental.jvm.StandaloneJavaApplicationPlugin"
+            implementationClass = "org.gradle.api.experimental.java.StandaloneJavaApplicationPlugin"
         }
         create("jvm-ecosystem") {
             id = "org.gradle.experimental.jvm-ecosystem"
