+++ unified-prototype/unified-plugin/plugin-jvm/src/main/java/org/gradle/api/experimental/java/JavaApplication.java
@@ -1,4 +1,4 @@
-package org.gradle.api.experimental.jvm;
+package org.gradle.api.experimental.java;
 
 import org.gradle.api.provider.Property;
 import org.gradle.declarative.dsl.model.annotations.Restricted;
